Refactor: 제품 등록 로직 제거로 인한 리팩토링, 함수제어자 추가

+++ contract/contracts/lib/Structs.sol
@@ -4,7 +4,7 @@ pragma solidity >=0.8.0 < 0.9.0;
 library Structs{
 
   struct Product {
-    uint32 id;
+    uint64 id;
     uint256 price;
     uint32 sellerId;
     address seller;
